+++ include/hermes_shm/memory/allocator/thread_local_allocator.h
@@ -73,7 +73,9 @@ struct _ThreadLocalAllocatorHeader : public AllocatorHeader {
   void FreeTid(hshm::ThreadId tid) { free_tids_->emplace(tid.tid_); }
 
   HSHM_INLINE_CROSS_FUN
-  TLS *GetTls(hshm::ThreadId tid) { return &(*tls_)[tid.tid_].tls_info_; }
+  TLS *GetTls(hshm::ThreadId tid) {
+    return &(*tls_)[(size_t)tid.tid_].tls_info_;
+  }
 };
 
 class _ThreadLocalAllocator : public Allocator {
@@ -166,7 +168,7 @@ class _ThreadLocalAllocator : public Allocator {
 
     // Case 1: Can we re-use an existing page?
     ThreadId tid = GetOrCreateTid(ctx);
-    PageAllocator &page_alloc = (*header_->tls_)[tid.tid_];
+    PageAllocator &page_alloc = (*header_->tls_)[(size_t)tid.tid_];
     page = page_alloc.Allocate(page_id);
 
     auto *tls_again = HERMES_THREAD_MODEL->GetTls<TLS>(tls_key_);
@@ -245,7 +247,7 @@ class _ThreadLocalAllocator : public Allocator {
     }
     hdr->UnsetAllocated();
     header_->SubSize(hdr->page_size_);
-    PageAllocator &page_alloc = (*header_->tls_)[hdr->tid_.tid_];
+    PageAllocator &page_alloc = (*header_->tls_)[(size_t)hdr->tid_.tid_];
     page_alloc.Free(hdr_offset, hdr);
   }
